added crud ep's

+++ addac03e2ac2f62d1a7e766874af16b3/Cities/City.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[City]
 (
-	[Id] INT IDENTITY(1, 1) PRIMARY KEY,
+	[Id] BIGINT IDENTITY(1, 1) PRIMARY KEY,
     [Name] NVARCHAR(50) NULL, 
     [State] NVARCHAR(50) NULL, 
     [Country] NVARCHAR(50) NULL,
